--- server/lib/ai/tools/index.ts
@@ -1,4 +1,0 @@
-export { createGetChannelMessagesTool } from "./get-channel-messages";
-export { createGetThreadMessagesTool } from "./get-thread-messages";
-export { joinChannelTool } from "./join-channel";
-export { searchChannelsTool } from "./search-channels";
